Active challenge page backend functionality implemented. OBS. Big change to int vs doubles.

+++ lib/pages/otherteampage.dart
@@ -1,14 +1,14 @@
 import 'dart:async';
 import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_application/api_utils/api_utils.dart';
+import 'package:flutter_application/components/custom_app_bar.dart';
 import 'package:flutter_application/components/custom_colors.dart';
 import 'package:flutter_application/components/donation_progress_bar.dart';
+import 'package:flutter_application/components/other_goal_box.dart';
 import 'package:flutter_application/components/return_arrow_argument.dart';
 import 'package:flutter_application/models/team.dart';
-import 'package:iconsax/iconsax.dart';
-import 'package:flutter_application/components/custom_app_bar.dart';
-import 'package:flutter_application/components/other_goal_box.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
 class OtherTeamPage extends StatefulWidget {
@@ -25,8 +25,8 @@ class _OtherTeamPageState extends State<OtherTeamPage> {
   String? teamName;
   String? companyName;
   String? charityName;
-  int donationGoal = 0;
-  int totalDonations = 0;
+  double donationGoal = 0;
+  double totalDonations = 0;
   late Timer timer;
   int teamRank = -1;
   int totalTeams = 0;
@@ -47,8 +47,8 @@ class _OtherTeamPageState extends State<OtherTeamPage> {
   Future<void> fetchTeamDetails() async {
     try {
       String? charity = await ApiUtils.fetchOtherCharityOrganization(teamName!);
-      int? goal = await ApiUtils.fetchOtherDonationGoal(teamName!);
-      int? donations = await ApiUtils.fetchOtherFundraiserBox(teamName!);
+      double? goal = await ApiUtils.fetchOtherDonationGoal(teamName!);
+      double? donations = await ApiUtils.fetchOtherFundraiserBox(teamName!);
       List<dynamic>? teamMembers = await ApiUtils.fetchOtherMembers(teamName!);
 
       setState(() {
@@ -105,10 +105,11 @@ class _OtherTeamPageState extends State<OtherTeamPage> {
     }
   }
 
+  // Används ej??
   Future<void> fetchDonationGoal() async {
     if (teamName != null) {
       try {
-        int? goal = await ApiUtils.fetchOtherDonationGoal(teamName!);
+        double? goal = await ApiUtils.fetchOtherDonationGoal(teamName!);
         setState(() {
           donationGoal = goal ?? 0;
         });
@@ -118,10 +119,11 @@ class _OtherTeamPageState extends State<OtherTeamPage> {
     }
   }
 
+  // Används ej??
   Future<void> fetchDonatedAmount() async {
     if (teamName != null) {
       try {
-        int? donations = await ApiUtils.fetchOtherFundraiserBox(teamName!);
+        double? donations = await ApiUtils.fetchOtherFundraiserBox(teamName!);
         setState(() {
           totalDonations = donations ?? 0;
         });
